--- src/main/java/com/example/SchedulingPro/Details/CustomOauth2UserDetails.java
@@ -1,79 +1,0 @@
-package com.example.SchedulingPro.Details;
-
-import com.example.SchedulingPro.User.User;
-import lombok.Getter;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Map;
-
-@Slf4j
-@Getter
-public class CustomOauth2UserDetails implements UserDetails, OAuth2User {
-
-    private final User user;
-    private Map<String, Object> attributes; // 소셜에서 받아온 정보들
-
-    public CustomOauth2UserDetails(User user, Map<String, Object> attributes) {
-        this.user = user;
-        this.attributes = attributes;
-    }
-
-    @Override
-    public Map<String, Object> getAttributes() {
-        return attributes;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        Collection<GrantedAuthority> collection = new ArrayList<>();
-        collection.add(new GrantedAuthority() {
-            @Override
-            public String getAuthority() {
-                return user.getRole();
-            }
-        });
-        return collection;
-    }
-
-    @Override
-    public String getPassword() {
-        return user.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return user.getUsername();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-
-    public String getName() {
-        return user.getName();
-    }
-
-    public String getEmail() {
-        return user.getEmail();
-    }
-}
