--- src/main/java/com/example/SchedulingPro/Details/KakaoUserDetails.java
@@ -1,31 +1,0 @@
-package com.example.SchedulingPro.Details;
-
-import lombok.AllArgsConstructor;
-
-import java.util.Map;
-
-@AllArgsConstructor
-public class KakaoUserDetails implements OAuth2UserInfo {
-
-    private Map<String, Object> attributes;
-
-    @Override
-    public String getCategory() {
-        return "kakao";
-    }
-
-    @Override
-    public String getCategoryId() {
-        return attributes.get("id").toString();
-    }
-
-    @Override
-    public String getEmail() {
-        return (String) ((Map) attributes.get("kakao_account")).get("email");
-    }
-
-    @Override
-    public String getName() {
-        return (String) ((Map) attributes.get("properties")).get("nickname");
-    }
-}
