--- src/main/java/com/example/SchedulingPro/Details/OAuth2UserInfo.java
@@ -1,13 +1,0 @@
-package com.example.SchedulingPro.Details;
-
-public interface OAuth2UserInfo {
-
-    String getCategory();
-
-    String getCategoryId();
-
-    String getEmail();
-
-    String getName();
-
-}
