--- src/main/java/com/example/SchedulingPro/Entity/Refresh.java
@@ -1,29 +1,0 @@
-package com.example.SchedulingPro.Entity;
-
-import jakarta.persistence.*;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Table(name = "refresh")
-@Entity
-public class Refresh extends BaseTime {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String username;
-
-    @Column(length = 512)
-    private String refresh;
-
-    private String expiration;
-
-    @Builder
-    public Refresh(String username, String refresh, String expiration) {
-        this.username = username;
-        this.refresh = refresh;
-        this.expiration = expiration;
-    }
-}
