--- src/main/java/com/example/SchedulingPro/Entity/RefreshRepository.java
@@ -1,13 +1,0 @@
-package com.example.SchedulingPro.Entity;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.transaction.annotation.Transactional;
-
-public interface RefreshRepository extends JpaRepository<Refresh, Long> {
-
-    Boolean existsByRefresh(String refresh);
-
-    @Transactional
-    void deleteByRefresh(String refresh);
-
-}
