--- src/main/java/com/example/SchedulingPro/User/UserRepository.java
@@ -1,8 +1,0 @@
-package com.example.SchedulingPro.User;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    User findByUsername(String username);
-}
