--- src/main/java/com/example/SchedulingPro/User/UserService.java
@@ -1,13 +1,0 @@
-package com.example.SchedulingPro.User;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-
-@Slf4j
-@Getter
-@RequiredArgsConstructor
-@Service
-public class UserService {
-}
