+++ src/main/java/com/example/SchedulingPro/config/WebConfig.java
@@ -1,0 +1,21 @@
+package com.example.SchedulingPro.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+    private static final String FRONT_SERVER = "http://localhost:3000";
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+                .allowedOrigins(FRONT_SERVER) // 이 주소에 권한 부여
+                .allowedMethods("GET", "POST", "PUT", "DELETE")
+                .exposedHeaders("location")
+                .allowedHeaders("*")
+                .allowCredentials(true); // true Cors 해결
+    }
+}
