+++ src/main/java/com/example/SchedulingPro/details/NaverUserDetails.java
@@ -1,0 +1,31 @@
+package com.example.SchedulingPro.details;
+
+import lombok.AllArgsConstructor;
+
+import java.util.Map;
+
+@AllArgsConstructor
+public class NaverUserDetails implements OAuth2UserInfo {
+
+    private Map<String, Object> attributes;
+
+    @Override
+    public String getCategory() {
+        return "naver";
+    }
+
+    @Override
+    public String getCategoryId() {
+        return (String) ((Map) attributes.get("response")).get("id");
+    }
+
+    @Override
+    public String getEmail() {
+        return (String) ((Map) attributes.get("response")).get("email");
+    }
+
+    @Override
+    public String getName() {
+        return (String) ((Map) attributes.get("response")).get("name");
+    }
+}
