+++ src/main/java/com/example/SchedulingPro/details/OAuth2UserInfo.java
@@ -1,0 +1,13 @@
+package com.example.SchedulingPro.details;
+
+public interface OAuth2UserInfo {
+
+    String getCategory();
+
+    String getCategoryId();
+
+    String getEmail();
+
+    String getName();
+
+}
