+++ src/main/java/com/example/SchedulingPro/entity/RefreshRepository.java
@@ -1,0 +1,13 @@
+package com.example.SchedulingPro.entity;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.transaction.annotation.Transactional;
+
+public interface RefreshRepository extends JpaRepository<Refresh, Long> {
+
+    Boolean existsByRefresh(String refresh);
+
+    @Transactional
+    void deleteByRefresh(String refresh);
+
+}
