+++ src/main/java/com/example/SchedulingPro/jwt/JwtFilter.java
@@ -1,0 +1,75 @@
+package com.example.SchedulingPro.jwt;
+
+import com.example.SchedulingPro.details.CustomOauth2UserDetails;
+import com.example.SchedulingPro.user.User;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.web.filter.OncePerRequestFilter;
+import java.io.IOException;
+
+@Slf4j
+@Setter
+@RequiredArgsConstructor
+public class JwtFilter extends OncePerRequestFilter {
+
+    private final JwtUtil jwtUtil;
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+
+        String accessToken = null; // accessToken 값
+        String refreshToken = null;
+
+        Cookie[] cookies = request.getCookies();
+
+        for (Cookie cookie : cookies) {
+
+            if (cookie.getName().equals("accessToken")) {
+                accessToken = cookie.getValue();
+            }
+
+            if (cookie.getName().equals("refreshToken")) {
+                refreshToken = cookie.getValue();
+            }
+        }
+
+        if (accessToken == null) {
+
+            log.info("token null");
+            filterChain.doFilter(request, response);
+
+            return;
+        }
+
+        // 만료시 다음 필터로 안넘김
+        if (jwtUtil.isExpired(accessToken)) {
+            log.info("token expired");
+            filterChain.doFilter(request, response);
+
+            return;
+        }
+
+        String username = jwtUtil.getUsername(accessToken);
+        String role = jwtUtil.getRole(accessToken);
+
+        User user = new User();
+        user.setUsername(username);
+        user.setRole(role);
+
+        CustomOauth2UserDetails customOauth2UserDetails = new CustomOauth2UserDetails(user, null);
+
+        Authentication authToken = new UsernamePasswordAuthenticationToken(customOauth2UserDetails, null, customOauth2UserDetails.getAuthorities());
+        SecurityContextHolder.getContext().setAuthentication(authToken);
+
+        filterChain.doFilter(request, response);
+    }
+}
