+++ src/main/java/com/example/SchedulingPro/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package com.example.SchedulingPro.repository;
+
+import com.example.SchedulingPro.user.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    User findByUsername(String username);
+}
