+++ src/main/java/com/example/SchedulingPro/schedule/ScheduleResponseDto.java
@@ -1,0 +1,16 @@
+package com.example.SchedulingPro.schedule;
+
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+public class ScheduleResponseDto {
+
+    private String title;
+    private String place;
+    private String start;
+    private String end;
+    private String content;
+
+}
