+++ src/main/java/com/example/SchedulingPro/service/UserService.java
@@ -1,0 +1,13 @@
+package com.example.SchedulingPro.service;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Slf4j
+@Getter
+@RequiredArgsConstructor
+@Service
+public class UserService {
+}
