+++ src/main/resources/application.properties
@@ -1,4 +1,6 @@
 spring.application.name=ChatProject
+spring.config.import=optional:file:.env[.properties]
+
 
 spring.datasource.driver-class-name = com.mysql.cj.jdbc.Driver
 spring.datasource.url = jdbc:mysql://127.0.0.1:3306/chat?useSSL=false&useUnicode=true&serverTimezone=Asia/Seoul
